\name{ProteinCharge-package}
\alias{ProteinCharge-package}
\alias{ProteinCharge}
\docType{package}
\title{
\packageTitle{ProteinCharge}
}
\description{
\packageDescription{ProteinCharge}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ProteinCharge}
\packageIndices{ProteinCharge}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{ProteinCharge}

Maintainer: \packageMaintainer{ProteinCharge}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
