% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titration.R
\name{titration}
\alias{titration}
\title{Calculate theoretical charge}
\usage{
titration(
  sequence = "",
  read_source = "",
  state = c("reduced", "oxidized"),
  chart_title = "",
  data_type = c("string", "file"),
  pH_start = 4,
  pH_stop = 10
)
}
\arguments{
\item{sequence}{The single letter abbreviated protein sequence. A string or character vector of length one.}

\item{read_source}{File path to the text file containing the protein sequence.}

\item{state}{whether to calculate the charge under oxidizing or reducing conditions. default is reducing.}

\item{chart_title}{The title to be displayed by ggplot2 above the produced graph}

\item{data_type}{Instructs function whether to consider the sequence or read_source arguments. default value is "string".}

\item{pH_start}{Starting point for theoretical titration. Must be numeric.}

\item{pH_stop}{End point of theoretical titration. Must be numeric}
}
\value{
Returns the protein charge to Stdout and creates graph in Plots panel
titration(sequence = "DYKDDDDK", chart_title = "Flag peptide")
titration(read_source = "flagPeptide.txt", chart_title= "Flag peptide", data_type = "file")
}
\description{
Calculate the theoretical charge of the given protein sequence over a range of pH values
}
